backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Redis
redis==5.0.1

# Document parsing
pdfplumber==0.10.3
python-pptx==0.6.23
python-docx==1.1.0
Pillow==10.1.0
pytesseract==0.3.10

# HTTP client for Ollama
httpx==0.25.2
aiohttp==3.9.1

# HTML parsing for schedule
beautifulsoup4==4.12.2

# CORS
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
